+++ src/exercise4.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 
 int main(int argc, char** argv) {
-    double x = atoi(argv[1]);
+    double x = atof(argv[1]);
     int k = atoi(argv[2]);
     double array[10];
 
@@ -11,12 +11,12 @@ int main(int argc, char** argv) {
     }
 
     for (int i = 0; i < k; i++) {
-        printf("%.1f ", array[i]);
+        printf("%.1lf ", array[i]);
     }
-    printf("%.1f ", x);
+    printf("%.1lf ", x);
 
     for (int i = k; i < 9; i++) {
-        printf("%.1f ", array[i]);
+        printf("%.1lf ", array[i]);
     }
 
     return 0;
